+++ lib/cache_x.dart
@@ -1,5 +1,6 @@
 library cache_x;
 
 import 'package:cache_x/src/cacheXWorker.dart';
+import 'package:flutter/material.dart';
 
 part 'src/cacheXCore.dart';