+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+
+- Updated storage implementation
+- Added named parameter
+
 ## 1.0.2
 
 - Updated Documentation